@@ -3,23 +3,20 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.5.0",
-    "@testing-library/user-event": "^7.2.1",
-    "clarifai": "^2.9.1",
-    "clarifai-nodejs-grpc": "^6.7.0",
-    "gh-pages": "^3.1.0",
-    "http2": "^3.3.7",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-particles-js": "^3.3.0",
-    "react-scripts": "3.4.3",
-    "react-tilt": "^0.1.4",
-    "serve": "^11.3.2",
+    "@testing-library/jest-dom": "^6.2.1",
+    "@testing-library/react": "^14.1.2",
+    "@testing-library/user-event": "^14.5.2",
+    "gh-pages": "^6.1.1",
+    "particles-bg": "^2.5.5",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-parallax-tilt": "^1.7.200",
+    "react-scripts": "^5.0.1",
+    "serve": "^14.2.1",
     "tachyons": "^4.12.0"
   },
   "scripts": {
-    "start": "serve -s build",
+    "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
