@@ -70,7 +70,7 @@ import React from 'react'
 			        onClick={this.onSubmitSignIn}
 			        className="b ph3 pv2 input-reset ba b--black bg-transparent grow pointer f6 dib"
 			        type="submit" 
-			        value="Register"/>
+			        value="Submit"/>
 			    </div>
 			  </div>
 			</main>
